refactor: DAM-8 Common으로 extension 이동
